test: with Jest and bug fix (#57)
* docs: update readme

#54

* feat: unit test using Jest and React Testing Library

#54

* fix: set static strings instead of env variables

#54

* fix: resolve the error of executing Jest

#54

* fix: resolve post undefined error by checking the type of post.identifier

#54

* feat: handling asynchronous status updates w/ waitFor and async/await
Resolve warning: An update to Home inside a test was not wrapped in act(...)
#54

* test: the act of deleting a sub

#54

* fix: add condition to check if postIds exist
Resolve DeleteSub.test.tsx FAIL
#54

* fix: Jest is not recognized by typescript
ref) https://stackoverflow.com/questions/72181689/jest-is-not-recognized-by-typescript
#54

* test: client side ai-response

#54

* fix: getAIResponse API bug

#54

* fix: clean up the codes

* feat: show 'Browse Other Communities' message when community does not exist

#54

* fix: failed to compile error on vercel
Error message:  can be escaped with , , , .  react/no-unescaped-entities
#54

* fix: update readme

+++ server/src/routes/subs.ts
@@ -90,26 +90,29 @@ const deleteSub = async (req: Request, res: Response) => {
             relations: ["posts", "posts.votes", "posts.comments"]
         });
         
-        // Delete post's votes
-        await Vote.createQueryBuilder()
-            .delete()
-            .from(Vote)
-            .where("postId IN (:...postIds)", { postIds: community.posts.map(post => post.id) })
-            .execute();
+        const postIds = community.posts.map(post => post.id);
+        if (postIds.length > 0) {
+            // Delete post's votes
+            await Vote.createQueryBuilder()
+                .delete()
+                .from(Vote)
+                .where("postId IN (:...postIds)", { postIds })
+                .execute();
 
-        // Delete post's comments
-        await Comment.createQueryBuilder()
-            .delete()
-            .from(Comment)
-            .where("postId IN (:...postIds)", { postIds: community.posts.map(post => post.id) })
-            .execute();
+            // Delete post's comments
+            await Comment.createQueryBuilder()
+                .delete()
+                .from(Comment)
+                .where("postId IN (:...postIds)", { postIds })
+                .execute();
 
-        // Delete posts
-        await Post.createQueryBuilder()
-            .delete()
-            .from(Post)
-            .where("subName = :subName", { subName: community.name })
-            .execute();
+            // Delete posts
+            await Post.createQueryBuilder()
+                .delete()
+                .from(Post)
+                .where("subName = :subName", { subName: community.name })
+                .execute();
+        }
 
         // Delete the sub
         const sub = await Sub.createQueryBuilder()
